+++ src/components/Fieldset/Fieldset.tsx
@@ -2,12 +2,18 @@ import classNames from "classnames";
 import PropTypes from "prop-types";
 import React from "react";
 
+import { Attributes, DataAttributes } from "../../helpers/Attributes";
+
 import "lbh-frontend/lbh/components/lbh-fieldset/_fieldset.scss";
 
 /**
  * The proptypes for {@link Fieldset}.
+ *
+ * This also supports all `aria-*` and `data-*` attributes.
+ *
+ * @noInheritDoc
  */
-export interface FieldsetProps {
+export interface FieldsetProps extends React.AriaAttributes, DataAttributes {
   id?: string;
   className?: string;
   role?: string;
@@ -27,11 +33,15 @@ export const Fieldset: React.FunctionComponent<FieldsetProps> = (
   props: FieldsetProps
 ): JSX.Element => {
   const { id, className, legend, role, describedBy, children } = props;
+
+  const extraAttributes = Attributes.ariaAndData(props);
+
   return (
     <fieldset
       id={id}
       className={classNames("govuk-fieldset lbh-fieldset", className)}
       role={role}
+      {...extraAttributes}
       aria-describedby={describedBy}
     >
       {legend}