+++ src/components/FormGroup.test.tsx
@@ -5,14 +5,22 @@ import { FormGroup } from "./FormGroup";
 
 it("renders correctly with all props", () => {
   const component = create(
-    <FormGroup id="Testid" className="class1 class2" error={true}>
+    <FormGroup
+      id="Testid"
+      className="class1 class2"
+      error={true}
+      aria-label="test form group"
+      data-test="test data"
+    >
       TestFormGroupChildren
     </FormGroup>
   );
 
   expect(component).toMatchInlineSnapshot(`
     <div
+      aria-label="test form group"
       className="govuk-form-group lbh-form-group class1 class2 govuk-form-group--error lbh-form-group--error"
+      data-test="test data"
       id="Testid"
     >
       TestFormGroupChildren