+++ src/components/icons/StartButtonIcon.tsx
@@ -1,10 +1,18 @@
 import PropTypes from "prop-types";
 import React from "react";
 
+import { Attributes, DataAttributes } from "../../helpers/Attributes";
+
 /**
  * The proptypes for {@link StartButtonIcon}.
+ *
+ * This also supports all `aria-*` and `data-*` attributes.
+ *
+ * @noInheritDoc
  */
-export interface StartButtonIconProps {
+export interface StartButtonIconProps
+  extends React.AriaAttributes,
+    DataAttributes {
   className?: string;
 }
 
@@ -15,7 +23,7 @@ export const StartButtonIcon: React.FunctionComponent<StartButtonIconProps> = (
   props: StartButtonIconProps
 ): JSX.Element => {
   const { className } = props;
-
+  const extraAttributes = Attributes.ariaAndData(props);
   // This SVG is sourced from `govuk-frontend@3.3.0`. See
   // https://github.com/alphagov/govuk-frontend/blob/c473c7ec3123e220d83221b9fd12e216f71bda84/package/govuk/components/button/template.njk#L27
   // for the source and
@@ -30,6 +38,7 @@ export const StartButtonIcon: React.FunctionComponent<StartButtonIconProps> = (
       viewBox="0 0 33 40"
       role="presentation"
       focusable="false"
+      {...extraAttributes}
     >
       <path fill="currentColor" d="M0 0h13l20 20-20 20H0l20-20z" />
     </svg>